+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "projet-j",
 	"private": true,
-	"version": "0.2.1a",
+	"version": "0.2.2a",
 	"type": "module",
 	"scripts": {
 		"dev": "npm-run-all --parallel dev:frontend dev:backend",
