+++ src/components/Header.tsx
@@ -1,23 +1,14 @@
-import { SignInButton, UserButton } from '@clerk/clerk-react'
-import { Authenticated, Unauthenticated } from 'convex/react'
-import img from '../images/Shutdown.svg'
+import Login from './Login'
 import Theme from './Theme'
 
 export default function Header() {
 	return (
-		<header className="h-16 bg-tone-ter text-tint-one">
+		<header className="h-16 bg-tone-bis text-tint">
 			<div className="h-16 container flex justify-between items-center">
 				<h1 className="text-2xl font-bold">PROJET J</h1>
 				<div className="flex gap-3 items-center">
 					<Theme />
-					<Authenticated>
-						<UserButton />
-					</Authenticated>
-					<Unauthenticated>
-						<SignInButton mode="modal">
-							<img src={img} alt="Login Button" className="w-8 h-8 p-1 flex cursor-pointer" />
-						</SignInButton>
-					</Unauthenticated>
+					<Login />
 				</div>
 			</div>
 		</header>
