--- src/components/styles/themeStyle.tsx
@@ -1,6 +1,0 @@
-const style = {
-	btn: 'h-8 w-8 m-0 p-2 flex',
-	btnGrp: 'bg-tone-ter text-tint-one flex h-8 w-24 p-0'
-}
-
-export default style
