+++ src/styles/themeStyle.tsx
@@ -1,0 +1,6 @@
+const style = {
+	btnGrp: 'bg-tone-ter text-tint flex p-0 items-center rounded-lg',
+	btn: 'p-1'
+}
+
+export default style
